Extract actual function acquisition
The pointers to the actual function implementations are now fetched inside the `actual` namespace declared in the `actual_function.h` header.

Fixed source of _noreturn_ related warning during compilation by adding the appropriate flag. Sadly this means that we can not use `std::function` in this context as it doesn't seem to carry these _c-like_ flags.